Move note helpers into utils package
