+++ src/local_server.py
@@ -49,7 +49,6 @@ from .models import (
     TTSProviders,
 )
 from .note_proccessor import NoteProcessor
-from .notes import get_fields, get_note_type_id_from_name
 from .prompt_helpers import get_extras, get_prompts_for_note
 from .sentry import sentry
 from .smart_field_models import (
@@ -62,6 +61,8 @@ from .smart_field_models import (
 from .smart_field_prompt_map import list_prompt_map, replace_from_prompt_map
 from .smart_field_service import smart_field_service
 from .ui.prompt_dialog import PromptDialog
+from .utils import get_fields
+from .utils.notes_utils import get_note_type_id_from_name
 
 FIELD_PATTERN = r"\{\{(?!c\d+::)(.+?)\}\}"
