+++ src/ui/addon_options_dialog.py
@@ -53,13 +53,13 @@ from ..decks import deck_id_to_name_map, deck_name_to_id_map
 from ..logger import logger
 from ..models import PromptMap, SmartFieldType, legacy_openai_chat_models
 from ..note_proccessor import NoteProcessor
-from ..notes import get_note_type_id_from_name
 from ..prompt_helpers import get_all_prompts, get_extras, get_prompts_for_note
 from ..smart_field_prompt_map import list_prompt_map, replace_from_prompt_map
 from ..smart_field_service import smart_field_service
 from ..tasks import run_async_in_background
 from ..telemetry import track_event
 from ..utils import get_fields, get_version
+from ..utils.notes_utils import get_note_type_id_from_name
 from .account_options import AccountOptions
 from .chat_options import ChatOptions
 from .image_options import ImageOptions
